fix: Se fixeo la forma que cambiaba a modo game despues de empezar el thread. Eso trae una race condition con el disconnect del player container. Y tambien esto conlleva cambios en la forma de saber si empezo o no. En el notifier

+++ tests/integration/testerclient.cpp
@@ -38,7 +38,8 @@ void TesterClient::assertLobbyStarted(uint8_t count) {
     lobby_info info;
     client.recvlobbyinfo(info);
 
-    EXPECT_EQ(info.action, LobbyResponseType::STARTED_LOBBY) << "INITED Match succesfully ";
+    EXPECT_EQ(info.action, (uint8_t)LobbyResponseType::STARTED_LOBBY)
+            << "INITED Match succesfully ";
     EXPECT_EQ(info.data, count) << "Total count at start is correct";
 }
